Fix cd error in test (#21)
Fix cd to runfiles when run via `bazel run`

The test runner is assuming it's being run via bazel-bin/path/to/test,
and immediately attempting to cd to $0.runfiles. However, if it is
being run via `bazel run` (or `bazel test`) bazel will essentially
execute `cd bazel-bin/path/to/target.runfiles; path/to/target`. If the
runner then attempts to cd to $0.runfiles, it will be attempting to
cd to the non-existent directory
bazel-bin/path/to/target.runfiles/path/to/target.runfiles.