Hard-code / as the flag start (#49)
This should work cross-platform (mono and otherwise.) This is important
for non-mono support on Windows.

https://manpages.debian.org/wheezy/mono-mcs/mcs.1.en.html

"The Mono C# compiler accepts the same command line options that the
Microsoft C# compiler does. Those options can start with a slash or a
dash (/checked is the same as -checked)."

This was tested on Linux and Mac with Mono.