Use assembly name for DotnetAssembly{Compile,Runtime}Info (#435)
Previously, we used the name of the target directly, which caused
problems in the following configuration:

    csharp_binary(
        name = "bar",
        out = "foo",
        // ...
    )

    publish_binary(
        name = "Bar",
        binary = ":bar",
    )

publish_binary uses the name specified in DotnetAssemblyRuntimeInfo to
generate the `{assembly}.deps.json` and `{assembly}.runtimeinfo.json` files,
which are required for the .NET application to run. Since this
was set to the target name, the generated filenames in the case above
would be `bar.deps.json` and `bar.runtimeinfo.json`, causing the
following error when trying to run the publish target:

    A fatal error was encountered. The library 'libhostpolicy.so'
    required to execute the application was not found in
    '/dotnet/sdk/path'.

The default case is unaffected, since assembly_name defaults to the
`name` attribute if `out` is not given.