csharp_assembly: Do not expose analyzer compile-/runtime deps to users (#482)
If an is_analyzer=True assembly is passed in as a dependency to another
target, we currently add its dependencies as if they're required at
runtime, even though the analyzer runs strictly at compile time. With
this change, the analyzer no longer contributes its dependencies to the
compile- or runtime dependency graph.