Fix handling of native runtime dependencies (#515)
## What?
Fix handling of native dependencies on `csharp/fsharp_binary` rules and in `publish_binary` when the publish is not self contained.

The .Net runtime expects native dependencies to be laid out in a specific folder structure so we create symlinks in the `csharp/fsharp_binary` and then use the correct folder structure when copying the files in `publish_binary`.